🐛 bugfix : lambda 코드 수정

+++ src/dynamoDB/connect.dynamodb.js
@@ -17,16 +17,17 @@ function lookupFunc() {
         camelData.forEach((event) => {
           eventNotificationHandler(event);
 
-          /* DB에 TTL 설정
           params.Key = { id: event.id };
-          docClient.delete(params, (err, data) => {
-            if (!err) {
-              console.log('DynamoDB 데이터 삭제 성공!', data);
-            } else {
-              console.log('DynamoDB 데이터 삭제 실패!', err);
-            }
-          });
-          */
+          const now = Date.now();
+          if (now - event.created >= 70000) {
+            docClient.delete(params, (err, data) => {
+              if (!err) {
+                console.log('DynamoDB 데이터 삭제 성공!', data);
+              } else {
+                console.log('DynamoDB 데이터 삭제 실패!', err);
+              }
+            });
+          }
         });
       }
     } else {
